+++ bigWig/src/jkweb/lib/sufx.c
@@ -35,7 +35,7 @@ if (h.majorVersion > SUFX_MAJOR_VERSION)
 	     "%s is version %d.",  fileName, SUFX_MAJOR_VERSION, fileName, h.majorVersion);
 
 struct sufx *sufx;
-verbose(2, "sufx file %s size %lld\n", fileName, h.size);
+verbose(2, "sufx file %s size %"PRIu64"\n", fileName, h.size);
 
 /* Get a pointer to data in memory, via memory map, or allocation and read. */
 struct sufxFileHeader *header ;
@@ -84,7 +84,7 @@ bits32 *chromSizes = sufx->chromSizes
 	= pointerOffset(header, mapOffset);
 mapOffset += sizeof(bits32) * chromCount;
 
-verbose(2, "total dna size %lld in %d chromosomes\n", (long long)header->dnaDiskSize, header->chromCount);
+verbose(2, "total dna size %"PRIdMAX" in %d chromosomes\n", (intmax_t)header->dnaDiskSize, header->chromCount);
 sufx->allDna = pointerOffset(header, mapOffset);
 mapOffset += header->dnaDiskSize;
